Allow user to specify name of study and provide output file download
We're using the react-jsonschema-form package to build a form from part
of the study JSON schema.

We've disabled native HTML validation and disabled live validation,
because we think it provides a better user experience and native HTML
validation components are not styleable.

We're using the download attribute on an anchor tag to encourage the
browser to download the JSON output file rather than render it in the
browser. This may not work in Internet Explorer, but we can address that
later if it's an issue.