Generate UUID field from name
The schema suggests the uuid field should be an md5 hash of the
name. As this needs to be present for the data to be valid, we've
decided to add a hidden, read-only form field whose value is generated
on-the-fly from the name field. We've needed to make `Form` a
"controlled component" to do this (i.e. passing `formData` into the
component). Since we can no longer rely on the presence of formData to
decide whether the form has been submitted, we've introduced a new
`submitted` state.

Co-authored-by: James Mead <james@floehopper.org>