Use study.json schema instead of constructing it
I'm deleting the non-required fields so that the behaviour doesn't change.

Note that I have to delete the `$schema` property so that validation works
as expected. I find this very strange. I've not looked very hard but
haven't found anything in the GitHub issues for react-jsonschema-forms[1]
that describes this problem.

[1]: https://github.com/rjsf-team/react-jsonschema-form/issues