Use official octokit JS library instead
Instead of the third party github-api we were using. We noticed that the
author of the library we were using has suggested it should be
deprecated[1], and we've also run into a few problems using it.

In switching we've also (we hope) solved the 422 errors we were seeing
when saving to GitHub.

Co-authored-by: Chris Lowis <chris.lowis@gmail.com>

[1]: https://github.com/github-tools/github/issues/564