Update schema files
By running `scripts/update-schema`.

Changes to the study schema mean that we now require a group to be
specified. I've updated the test to enter the minimum information
required to create a group as part of the study. I've also updated the
Group property titles so that I can target them more specifically in the
tests. I've also set Group Inclusion criteria minItems to 1 so that the
form contains a single inclusion textbox on first render. I _think_ this
matches the intention of including "inclusion" in the array of required
properties.