Dereference schema files before use
We now pipe the raw schema file from GitHub through the dereference-schema
script before prettifying it. This means that we no longer have to
manually dereference the Person schema in the Datasets and Study
components.

This should make it easier to include the Device schema as that also
requires dereferencing against external schema.