Include contributors in Study schema
I'm doing this by including the Person schema in our project and
manually manipulating the Study schema to use this Person schema.

I've updated the test to enter information in the additional required
fields.

I also experimented with using the json-schema-ref-parser library[1] to
automatically dereference the link to the external Person schema but
this results in an HTTP request being made on page load to retrieve the
Person schema - which also means our App.test requires an HTTP request
too.

I think an improvement might be to add a pre-processor type script that
uses json-schema-ref-parser to dereference the schema before we then use
them in this app. I've had a quick play in that direction but it feels
like a distraction just now.

[1]: https://github.com/APIDevTools/json-schema-ref-parser