Update schema to get datasets nested in study
Schema updated using ./scripts/update-schema.

Having datasets nested in study avoids us having to construct this
relationship in memory.

I've kept the interface the same by removing the datasets property in
the Study component, and by extracting it from the study schema in the
Datasets component. I initially tried deleting the `properties.datasets`
key in the Study component but that meant the property was unavailable
for use in the Datasets component.