Update schema to get events nested in participant
Schema updated using ./scripts/update-schema.

Having events nested in participant avoids us having to construct this
relationship in memory.

I've kept the interface the same by removing the events property in the
Participant component, and by extracting it from the study schema in the
Events component. I might change this so that we do nest Events in the
Participant in the interface, though.