Add script to validate data against study schema
This highlights a problem with the data we're currently generating: it
can't be validated against the study schema because it's really some
kind of proxy-project. I'm going to look at getting the project schema
in so that we can validate our generated data against that.

We have to use ajv-formats-draft2019 to get support for idn-email.